Clean up - Update playwright configs for default folders - Move App.spec.tsx and screenshots - Update README.md for example - Revert changes to App.tsx to demonstrate snapshot - Remove headless false option
