+++ sugar/react/example/tests/App.spec.tsx
@@ -1,10 +1,8 @@
 import { expect, test } from "@playwright/experimental-ct-react";
 
-import { App } from "./App";
+import { App } from "../App";
 
-// test.use({ viewport: { width: 500, height: 500 } });
-
-test("should work", async ({ mount }) => {
+test("should match snapshot", async ({ mount }) => {
   const component = await mount(<App />);
   await expect(component).toHaveScreenshot();
 });
